Add files via upload
dirmakr.sh will create an organized directory system and run simulations in them. $cx is the parameter for the the heterochromatin-lamina interaction strength (0.0 = no interaction, 0.2 = average interaction, 0.4 = strong interaction). The #SBATCH lines can be removed or filled out depending on if you choose to run your simulations using slurm or not.