Add files via upload
radialdensityradii.py will obtain the radii for each chromatin subtype within the system and write them to several text files. These text files will be used as inputs for the radialdensitycalc.py which will histogram the data and perform the additional calculations necessary for obtaining the radial densities for each chromatin subtype. It will also generate a new set of files specifically for plotting the radial densities with the radialdensityplot.py script.